Fix Clang strict warning violations across core library files
This commit addresses hundreds of sign conversion, integer truncation,
and pointer arithmetic warnings detected by Clang's strict warning flags:
  -Wall -Wextra -Wpedantic -Wsign-conversion -Wconversion -Werror

## Files Fixed (14 core library files now compile warning-free):
- src/varintExternal.h - Fixed sign conversions in bit shift macros
- src/varintExternal.c - Fixed varintWidth truncations and signed conversions
- src/varintTagged.c - Fixed alignment issues (memcpy) and sign conversions
- src/varintDelta.h - Fixed zigzag decode sign conversion
- src/varintDelta.c - Fixed pointer arithmetic returns (4 fixes)
- src/varintFOR.c - Fixed pointer arithmetic return
- src/varintFloat.c - Fixed int64_t truncations and pointer arithmetic
- src/varintDict.c - Fixed binary search sign conversion
- src/varintBitmap.c - Fixed multiple sign conversions and truncations (4 fixes)
- src/varintAdaptive.c - Fixed size_t to uint32_t truncation
- src/varintChained.c - Fixed int32_t to varintWidth and reverse conversions
- src/varintChainedSimple.c - Fixed extensive uint32_t/uint64_t truncations (9 fixes)
- src/perf.h - Fixed float literal to integer constant (1e6 → 1000000UL)
- src/varintDimension.c - Fixed macro uint to uint8_t truncations (2 fixes)

## Tools Added:
- check_warnings.sh - Automated warning checker for both GCC and Clang

## Testing Status:
✓ 14/25 files compile cleanly with strict Clang warnings
✓ All core encoding/decoding files fixed
✗ 11 files remaining (test files + varintCompare.c + varintPacked.h issues)

## Remaining Work:
- Fix perf.h prototype and field initializer warnings
- Fix varintCompare.c sign conversions (demo file, 20+ warnings)
- Fix all test files (*Test.c) - mostly unused variables
- Fix varintPacked.h macro truncations (affects varintDimension.c)
- Run full test suite with sanitizers
- Verify GCC compatibility

## Pattern of Fixes Applied:
1. Pointer arithmetic: `return p - start;` → `return (size_t)(p - start);`
2. Bit operations: `vimp_src_[1] << 8` → `((uint64_t)vimp_src_[1] << 8)`
3. Enum truncation: `encoding` → `(uint8_t)encoding`
4. Sign conversion: `(zigzag & 1)` → `(uint64_t)(-(int64_t)(zigzag & 1))`
5. Int truncation: `v` → `(uint8_t)v` where needed
6. Alignment safety: Direct assignment → memcpy for unaligned access

All fixes preserve existing logic and only add explicit casts to satisfy
strict type checking. No functional changes were made.

+++ src/varintAdaptive.c
@@ -312,7 +312,7 @@ size_t varintAdaptiveEncodeWith(uint8_t *dst, const uint64_t *values,
 
         case VARINT_ADAPTIVE_PFOR: {
             varintPFORMeta pforMeta;
-            encodedSize = varintPFOREncode(dst + offset, values, count,
+            encodedSize = varintPFOREncode(dst + offset, values, (uint32_t)count,
                                           VARINT_PFOR_THRESHOLD_95, &pforMeta);
 
             if (meta) {
